pyproject.toml
[project]
name = "dio_image_processing"
version = "0.1.0"
description = "A simple image processing package for combining images."
authors = [
    {name = "Kelven Silva",email = "kelvenpersilva@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ultralytics (>=8.3.145,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
